Code restructuring
1 - Simplification of orchestors.
2 - Reorganization of functions.
3 - Added parsing of multiples inputs with and without quotes.
4 - Modified the atoi function to only convert characters to integers
  whithout handling errors.
5 - Added validations in check_entry.c

+++ src/linked_list_utils.c
@@ -6,13 +6,13 @@
 /*   By: lpin <lpin@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/28 19:01:58 by lpin              #+#    #+#             */
-/*   Updated: 2024/10/03 20:58:13 by lpin             ###   ########.fr       */
+/*   Updated: 2024/10/04 19:54:16 by lpin             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../header/push_swap.h"
 
-t_ps	*ft_ps_new(int content)
+t_ps	*ft_ps_new(long content)
 {
 	t_ps	*nodo;
